Fix ESLint configuration
Now that the project's package.json file includes `"type": "module"`,
ESLint no longer works with the default .eslintrc.js file. I tried
changing the syntax of this file to `export default`, but that doesn't
seem to work either. Instead, the simple fix is to use a CommonJS module
for now.